+++ idempotence4j-core/src/main/java/com/transferwise/idempotence4j/core/ActionRepository.java
@@ -8,6 +8,22 @@ public interface ActionRepository {
     Optional<Action> find(ActionId actionId);
     Action insertOrGet(Action action);
     Action update(Action action);
-    void deleteOlderThan(Instant timestamp, Integer batchSize);
-    void deleteByIds(List<ActionId> actionIds);
+
+    /**
+     * Deletes actions from the repository that were first executed before the provided timestamp, and limited by the batch size.
+     * @return the number of actions deleted
+     */
+    int deleteOlderThan(Instant timestamp, int batchSize);
+
+    /**
+     * Batch deletes actions from the repository by type and client, limited by the batch size.
+     * @return the number of actions deleted
+     */
+    int deleteByTypeAndClient(String type, String client, int batchSize);
+
+    /**
+     * Batch deletes actions from the repository by their IDs.
+     * @return An array containing the number of actions deleted, per action ID passed in as an argument.
+     */
+    int[] deleteByIds(List<ActionId> actionIds);
 }