+++ idempotence4j-mariadb/src/main/java/com/transferwise/idempotence4j/mariadb/JdbcMariaDbActionRepository.java
@@ -48,25 +48,33 @@ public class JdbcMariaDbActionRepository implements ActionRepository {
 	}
 
     @Override
-    public void deleteOlderThan(Instant timestamp, Integer batchSize) {
-        MapSqlParameterSource queryParameters = new MapSqlParameterSource()
+    public int deleteOlderThan(Instant timestamp, int batchSize) {
+        MapSqlParameterSource deleteParameters = new MapSqlParameterSource()
             .addValue("createdAt", Timestamp.from(timestamp))
             .addValue("limit", batchSize);
 
-        List<ActionId> actionIdList = namedParameterJdbcTemplate.query(FIND_OLDER_THAN_SQL, queryParameters, (rs, rowNum) -> sqlMapper.toId(rs));
-
-        deleteByIds(actionIdList);
+        return namedParameterJdbcTemplate.update(DELETE_OLDER_THAN_SQL, deleteParameters);
     }
 
     @Override
-    public void deleteByIds(List<ActionId> actionIdList) {
+    public int deleteByTypeAndClient(String type, String client, int batchSize) {
+        MapSqlParameterSource deleteParameters = new MapSqlParameterSource()
+            .addValue("type", type)
+            .addValue("client", client)
+            .addValue("limit", batchSize);
+
+        return namedParameterJdbcTemplate.update(DELETE_BY_TYPE_AND_CLIENT_SQL, deleteParameters);
+    }
+
+    @Override
+    public int[] deleteByIds(List<ActionId> actionIdList) {
         MapSqlParameterSource[] deleteBatchParameters = actionIdList.stream().map(actionId -> new MapSqlParameterSource()
             .addValue("key", actionId.getKey())
             .addValue("type", actionId.getType())
             .addValue("client", actionId.getClient()))
             .toArray(size -> new MapSqlParameterSource[size]);
 
-        namedParameterJdbcTemplate.batchUpdate(DELETE_SQL, deleteBatchParameters);
+        return namedParameterJdbcTemplate.batchUpdate(DELETE_BY_ACTION_ID_SQL, deleteBatchParameters);
     }
 
     //@formatter:off
@@ -119,15 +127,20 @@ public class JdbcMariaDbActionRepository implements ActionRepository {
 				"result = :result, " +
 				"result_type = :resultType";
 
-    private final static String FIND_OLDER_THAN_SQL =
-        "SELECT " +
-            "`key`, type, client " +
-            "FROM idempotent_action " +
+    private final static String DELETE_OLDER_THAN_SQL =
+        "DELETE FROM idempotent_action " +
             "WHERE " +
             "created_at < :createdAt " +
             "LIMIT :limit";
 
-    private final static String DELETE_SQL =
+    private final static String DELETE_BY_TYPE_AND_CLIENT_SQL =
+        "DELETE FROM idempotent_action " +
+            "WHERE type = :type " +
+            "  AND client = :client " +
+            "ORDER BY created_at ASC " +
+            "LIMIT :limit";
+
+    private final static String DELETE_BY_ACTION_ID_SQL =
         "DELETE " +
             "FROM idempotent_action " +
             "WHERE " +