+++ idempotence4j-mariadb/src/test-integration/groovy/com/transferwise/idempotence4j/mariadb/JdbcMariaDbActionRepositoryIntegrationTest.groovy
@@ -9,6 +9,8 @@ import spock.lang.Subject
 import java.time.Clock
 import java.time.Instant
 
+import static com.transferwise.idempotence4j.factory.ActionTestFactory.CLIENT
+import static com.transferwise.idempotence4j.factory.ActionTestFactory.TYPE
 import static com.transferwise.idempotence4j.factory.ActionTestFactory.anAction
 import static com.transferwise.idempotence4j.factory.ActionTestFactory.anActionId
 import static com.transferwise.idempotence4j.factory.ActionTestFactory.anActionResult
@@ -79,16 +81,43 @@ class JdbcMariaDbActionRepositoryIntegrationTest extends IntegrationTest {
         and:
             actions.each { it -> repository.insertOrGet(it) }
         when:
-            repository.deleteOlderThan(Instant.parse("2019-12-11T00:00:00Z"), 5)
+            int firstDeletionCount = repository.deleteOlderThan(Instant.parse("2019-12-11T00:00:00Z"), 5)
         then:
+            firstDeletionCount == 5
             countActions() == 15
         when:
-            repository.deleteOlderThan(Instant.parse("2019-12-11T00:00:00Z"), 5)
+            int secondDeletionCount = repository.deleteOlderThan(Instant.parse("2019-12-11T00:00:00Z"), 5)
         then:
+            secondDeletionCount == 5
             countActions() == 10
         and:
             purged.each { Action it ->
                 assert repository.find(it.actionId).isEmpty()
+            }
+    }
+
+    def "should remove actions by type and client"() {
+        given:
+            def actions = []
+            10.times {
+                actions << anAction()
+            }
+        and:
+            List<ActionId> actionIds = actions.each { it -> repository.insertOrGet(it) } .collect({ it.actionId })
+            def firstHalfActionIds = actionIds.dropRight(actions.size() / 2 as int)
+            def lastHalfActionIds = actionIds.drop(actions.size() / 2 as int)
+        when:
+            int firstDeletionCount = repository.deleteByTypeAndClient(TYPE, CLIENT, 5)
+        then:
+            firstDeletionCount == 5
+            firstHalfActionIds.each { ActionId it -> assert repository.find(it).isEmpty() }
+            lastHalfActionIds.each { ActionId it -> assert !repository.find(it).isEmpty() }
+        when:
+            int secondDeletionCount = repository.deleteByTypeAndClient(TYPE, CLIENT, 5)
+        then:
+            secondDeletionCount == 5
+            actionIds.each { ActionId it ->
+                assert repository.find(it).isEmpty()
             }
     }
 
@@ -103,12 +132,14 @@ class JdbcMariaDbActionRepositoryIntegrationTest extends IntegrationTest {
             def firstHalfActionIds = actionIds.dropRight(actions.size() / 2 as int)
             def lastHalfActionIds = actionIds.drop(actions.size() / 2 as int)
         when:
-            repository.deleteByIds(firstHalfActionIds)
+            int[] firstRowsDeleted = repository.deleteByIds(firstHalfActionIds)
+            firstRowsDeleted == [1, 1, 1, 1, 1] as int[]
         then:
             firstHalfActionIds.each { ActionId it -> assert repository.find(it).isEmpty() }
             lastHalfActionIds.each { ActionId it -> assert !repository.find(it).isEmpty() }
         when:
-            repository.deleteByIds(lastHalfActionIds)
+            int[] secondRowsDeleted = repository.deleteByIds(lastHalfActionIds)
+            secondRowsDeleted == [1, 1, 1, 1, 1] as int[]
         then:
             actionIds.each { ActionId it ->
                 assert repository.find(it).isEmpty()